Change osm_id to Long (#23)

+++ core/src/main/java/de/thb/core/data/sources/district/DistrictService.kt
@@ -17,7 +17,7 @@ interface DistrictService {
     @GET("/details.php")
     suspend fun getByOsm(
         @Query("osmtype") osmType: String,
-        @Query("osmid") osmId: Int,
+        @Query("osmid") osmId: Long,
         @Query("format") format: String,
     ): DetailsResponse
 }
